Fix button backlights that depend on ALS.
On a few devices, lighting the buttons when the screen turns on
depends on the lightsensor, regardless of the user preference.
So go back to enabling the sensor whenever the screen turns on, and
enforce the user preference after running through the light settings
once.

Port from CM9.

Original: http://r.cyanogenmod.com/#/c/12818/

Change-Id: Ibb9155b4f299637d553986c44b9cb548dc855489