SamsungRIL: Fix data connection after pin unlock
The source still has PUK code in it. For now this patch makes sense.

If the user has PIN-locked his phone, data connectivity will not
work because the RIL sends APPSTATE_UNKNOWN and does not update
it to APPSTATE_READY. We inject the state update if there's a
signal change.

Change-Id: Id64e95ebcadd1d7ff181f8d9c2ee4f18e6ae48f1